Fixes to elimiinate some ifdefs.
Moved UnzipIfNecessary to CommandProcessor so that UnZip can work on PerfCollect.
Removed some fidef DOTNET_CORE that are not needed in V2.0
