+++ src/PerfViewCollect/PerfViewCollect.csproj
@@ -15,6 +15,7 @@
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+    <NoWarn>1573,1591,0649</NoWarn>
   </PropertyGroup>
 
   <ItemGroup>
